+++ ThirdPartyAdapters/pangle/pangle/src/main/java/com/google/ads/mediation/pangle/PangleRequestHelper.java
@@ -1,3 +1,17 @@
+// Copyright 2023 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package com.google.ads.mediation.pangle;
 
 import android.text.TextUtils;
@@ -9,8 +23,8 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class PangleRequestHelper {
-    public static void fillWaterCoverParam(PAGRequest request, String bidResponse, MediationAdConfiguration adConfiguration) {
-        // only accept bidding scene
+    public static void setWatermarkString(PAGRequest request, String bidResponse, MediationAdConfiguration adConfiguration) {
+        // Only bidding ads require watermark.
         if (TextUtils.isEmpty(bidResponse)) {
             return;
         }
@@ -25,4 +39,6 @@ public class PangleRequestHelper {
         extraInfo.put("admob_watermark", watermark);
         request.setExtraInfo(extraInfo);
     }
+
+    private PangleRequestHelper() {}
 }